Move tests to top-level module
